requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.0
Flask-Migrate>=4.1.0
Werkzeug>=3.0.0
WTForms>=3.1.0
email-validator>=2.1.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
PyJWT>=2.9.0
SQLAlchemy>=2.0.25
alembic>=1.13.0
